base for VR talking action
Done: module that gets attached to a patient or similar object.
Module creates a "notification" prefab to display when it's possible to talk with the person (only during correct action)
Module allows to link dialogue xml files to load chat options from.
Completed a function to initialize selectionDialogue with proper text and functions.