Refactored coding of Action system
Moved all the variables into a single structure, so whenever we add new fields on the actions - we don't need to add them to dozens of different places (previously to every single constructor and it's use within ActionManager). Now it's only 2 places even if we add 10 more action types.
Tested on 4 different scenes, completing each from the beginning to the end few times - no issues detected.